fix: Switching currency in settings now changes the currency on home screen
