+++ DashWallet/Sources/Application/Constants.swift
@@ -18,3 +18,5 @@
 import Foundation
 
 let kOneDash: UInt64 = 100_000_000;
+
+
